move non-demo specific MPC related code to matchernet/control dir #13
