+++ demo/car_control/debug_display.py
@@ -3,7 +3,8 @@ import numpy as np
 import pygame, sys
 import cv2
 
-from mpcenv import MPCEnv
+from matchernet import MPCEnv
+
 from car import CarDynamics, CarRenderer
 from obstacle_reward_system import ObstacleRewardSystem
 from utils import MovieWriter
