+++ matchernet/__init__.py
@@ -1,4 +1,7 @@
 # -*- coding: utf-8 -*-
-__all__ = ["matchernet"]
+#__all__ = ["matchernet", "control"]
 
 from .matchernet import *
+from .control.mpc import *
+from .control.mpcenv_bundle import *
+from .control.mpcenv import *
